Neural Network rework

+++ XOR/Maths.c
@@ -6,9 +6,9 @@
 /// </summary>
 /// <param name="x"> : a double</param>
 /// <returns>the sigmoid of x</returns>
-float sigmoid(float x)
+double sigmoid(double x)
 {
-  return 1.0 / (1.0 + exp(-x));
+  return 1 / (1 + exp(-x));
 }
 
 /// <summary>
@@ -16,7 +16,7 @@ float sigmoid(float x)
 /// </summary>
 /// <param name="x">is already a sigmoid</param>
 /// <returns>the derivate of x</returns>
-float der_sigmoid(float x)
+double der_sigmoid(double x)
 {
 	return x * (1 - x);
 }
