+++ XOR/Maths.h
@@ -1,7 +1,7 @@
 #ifndef MATHS_H
 #define MATHS_H
 
-float sigmoid(float x);
-float der_sigmoid(float x);
+double sigmoid(double x);
+double der_sigmoid(double x);
 
 #endif
